Adjust arc z value for better visibility
Push back export arcs, so when the get large, they do not
obscure other parts of the graph.
Also, if an arc is part of a selected flow, ie its source or
destination node is selected, then push it forward to the viewer
slightly, so that non selected arcs do not obscure it.